Changed Double to not use "ushort", instead using size_t.
Changed utils to use sysconf to lookup the size of hostnames instead of HOST_NAME_MAX.

+++ src/Double.cpp
@@ -289,7 +289,7 @@ int Double::toInt()
 		int returnMe = 0;
 		if(hex.find(".") == hex.npos)
 			throw DavidException(hex+" is not a valid hexadecimal number",DavidException::FORMAT_ERROR_CODE);
-		for(ushort i = 0;i<hex.length();i++)
+		for(size_t i = 0;i<hex.length();i++)
 		{
 			try
 			{
